Set a more reasonable timeoutlen
With the default timeoutlen of 1000ms, it is very noticeable when you
use something like kk to escape while in insert mode. If you wanted to
type kk instead of using it to escaape, you have to wait a whole second.
It seems like 175ms is a nice choice to get around this very long delay.